(incremental-meta-1) Add Type Kind
Summary:
This is a series of diffs that begins the process of making
the meta information about definitions (and the types within) serializable.

The general approach is going to be make a series of intermediate objects
that are whitelisted for serdes. Along the way that has exposed some
oddities in the core type system that will also be improved.

Step one is to add an enum to encode all the different type kinds in the
config type system that encode behavior changes. Currently this was
only implemented via subclassing and overriding methods that return bool
or not. This is not a serializable format. I decided to add a type kind
within the config type system proper. I actually think this is more clear anyway

Test Plan: BK

Reviewers: max, alangenfeld

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D1467