Allow frozenset in set_param checks
Summary:
Where appropriate, I think we should start to use frozenset instead
of sets. Yesterday I had a nasty nasty bug where test order effected whether
or not a set passes. A reference to set was changed and effected a subsequent
run. I used a frozenset to debug the issue and would have prevented the bug
in the first place (as in the error would be made obvious)

Test Plan: BK

Reviewers: max, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1498