Refactor solid selector to defer query until selector is opened
Summary:
The query is loaded prior to opening the selector modal solely for the purpose of calculating the button label.  This diff instead calculates the label upon solid selection and persists it (along with the selection).

This diff has the effect that prior to going through the solid selection process, the label may change for people from something like `A -> C` to `3 solids`.

@bengotow: I kept the `fetchPolicy` as `network-only`, but I think it should probably be the default `cache-first`.  I wasn't able to repro any caching issues when changing variables, but wasn't confident enough to change it.  It would be nice to not have to incur this fetch every time you open the selector (used to be only on loading the `PipelineExecutionContainer`).

{F72475}

Test Plan: Loaded solid selector, switched pipelines, loaded again, forced spinner to show up

Reviewers: bengotow, schrockn, alangenfeld

Reviewed By: bengotow, alangenfeld

Subscribers: bengotow

Differential Revision: https://dagster.phacility.com/D1627