(new-config-parsing-5) TraversalContext --> ValidationContext
Summary:
And extract config_value out of ValidationContext.

This was odd to model in one's head IMO as the thing being operated
on was different than the context. The config value gets mutated
and the other items did not so this seems like a natural division

Depends on D1669

Test Plan: BK

Reviewers: nate, alangenfeld

Reviewed By: nate

Differential Revision: https://dagster.phacility.com/D1671