(new-config-parsing-6) More principled treatment of None in the config system no none
Summary:
This has always been pretty fast and loose and I think actually
causes some ambiguities and confusing behavior. Now the system has a principled
view of things. If it's Nullable or Any, None is valid. Otherwise, it is not.

The corner case this exposes wis when we are dealing with selectors.
There are cases where the presence of a selector key just by itself is
enough information to know what configuration should be. In this case
we do accept None.

Depends on D1671

Test Plan: BK, dagit

Reviewers: alangenfeld, nate, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1672