Make Runtime Types normal objects
Summary:
It's increasingly clear to me that typing Dagster Types with
Python Types in such a coupled way was a mistake.

And it turns out that with the type registy we don't need to be
dynamically generating classes anymore.

This diff is a proposal to, much like with the config types, move
away from the notion that runtime types to be python classes totally.

This diff still leaves some things unsolved. In particular what happens
when there isn't a clean bijection (1-to-1 relationship) between the
dagster runtime type domain and the python type domain.

@themissinghlink has been butting his head against this directly
while exploring customized data frame types. Meaning that we want
N types in the dagster domain for modeling different natures of
dataframes. But the *runtime* representation we want to be
of a single type, the pandas data frame.

Makes runtime types vanilla objects makes this separation clear.

However this still leaves open the question of what we want
to do about type annotations in a world where the bijection does
not exist.

I think there are a few syntaxes available to use, but in the short
term, if you want to write solids where the dagster types don't
1-to-1 map to python types, you just have to use InputDefinition
and OutputDefinition

This is mostly for discussion, not for check-in. I would clean up
the test cases in slightly different ways.

Test Plan: BK

Reviewers: max, alangenfeld, themissinghlink

Reviewed By: max, alangenfeld

Subscribers: themissinghlink

Differential Revision: https://dagster.phacility.com/D1685