(0/n run-storage) Update storage tags methods to take dictionary of tags instead of list of key values
Summary:
Previously, `get_run_count_with_matching_tags` and `get_runs_with_matching_tags` took in a list of `(key, value)` tuples. However, it makes more sense to take in a dictionary, to be consistent with how tags are represented throughout the rest of the codebase.

Note: this would be a breaking change

Test Plan: unit

Reviewers: alangenfeld, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D1787