[RFC] Eliminate get_runs_with... methods on RunStorage and DagsterInstance
Summary: With the introduction of the `get_runs` and `get_runs_count` methods, we can consolidate all of the `get_runs_with...` and `get_runs_count_with...` methods that exist on `RunStorage. This diff deletes all of the `get_runs_with...` and `get_runs_count_with...` methods and replaces calls to them with calls to `get_runs` and `get_runs_count` with the appropriate filters.

Test Plan: unit

Reviewers: prha, alangenfeld

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D1792