Replace playground subset selector with one that leverages solid query syntax
Summary:
This diff modifies the playground's solid subset selector to use the solid query syntax and the same rules for rendering / not-rendering (*), which fixes an issue encountered opening this modal on large DAGs.

It also fixes an issue where the DAG would render off in a corner of the modal, which was caused by the dialog's animation changing the SVGViewport's perceived size.

Under the hood, we're still converting the selection to a solid name array to pass to execution. It'd be nice to eliminate the solidSubset:[] param and use solidQuery everywhere, but we can implement the backend portion separately.

{F77471}

{F77472}

{F77473}

Test Plan: Run snapshot tests

Reviewers: alangenfeld, schrockn, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D1798