Fix errors with GCP pipeline
Summary: Some of the breaking changes, to the config system left the gcp pipeline in a broken state. This diff fixes the config errors. Specifically, the input defs of composite solids have to match the first solid's input defs, and the output defs of a composite solid have to match the last solid's output_defs.

Test Plan: Add test

Reviewers: max, schrockn, nate

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D1835