add collapsed tags view in RunTable
Summary:
This diff adds a condensed view of tags in RunsTable.  It drops the built-in `dagster`
prefix in favor of a subtle coloring of the tags to a different shade of blue (depending on the
key name).  Custom tags will remain gray.

It also by default limits the tag display to a single row, with an ellipsis for overflowing content
which expands to a multiline display upon a mouse click.  Instead of using cell measurement, which
is quite slow due to the number of rows involved, it hacks this together by always showing an
ellipsis which is then clipped by content.

{F80294}

Test Plan: Created a run with many tags, with a couple tags, with no tags.

Reviewers: bengotow, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1861