Combine start/launch buttons into a toggleable execute button group
Summary:
This diff groups the start and launch button (if they both are eligible) into
a button group, where a dropdown button opens a menu to select the execution type.  The
selection is stored in local storage as a global setting (sticky across sessions).

Implementation-wise, this diff changes `ExecutionStartButton` in order to extend from
BlueprintJS's `Button` component.  It can then use the contextual styling when the button
is wrapped with a `<ButtonGroup />` component. It uses a `Popover` component to implement
the menu and customizes the menu items to include a long-form description.

{F83257}

Test Plan: bk, tested with multiple connection states, with no run launcher, etc.

Reviewers: bengotow, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1871